提交

+++ tangtang/src/main/java/net/ttcxy/tangtang/service/impl/FansServiceImpl.java
@@ -1,12 +1,11 @@
 package net.ttcxy.tangtang.service.impl;
 
 import cn.hutool.core.util.IdUtil;
-import net.ttcxy.tangtang.entity.dto.Fans;
-import net.ttcxy.tangtang.entity.dto.User;
+import net.ttcxy.tangtang.entity.Fans;
+import net.ttcxy.tangtang.entity.User;
 import net.ttcxy.tangtang.mapper.FansMapper;
 import net.ttcxy.tangtang.mapper.UserMapper;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
@@ -29,6 +28,9 @@ public class FansServiceImpl {
 
     public Boolean selectFans(String fansName){
         Fans fans = getFans(fansName);
+        if (fans==null){
+            return false;
+        }
         int isOk = fansMapper.selectFans(fans);
         if (isOk == 1){
             return true;
@@ -36,17 +38,27 @@ public class FansServiceImpl {
         return false;
     }
 
-    public Boolean insertFans(String fansName) throws DuplicateKeyException {
-        Fans fans = getFans(fansName);
-        int isOk = fansMapper.insertFans(fans);
-        if (isOk == 1){
-            return true;
+    public Boolean insertFans(String fansName) {
+        try{
+            Fans fans = getFans(fansName);
+            if (fans==null){
+                return false;
+            }
+            int isOk = fansMapper.insertFans(fans);
+            if (isOk == 1){
+                return true;
+            }
+        }catch (Exception e){
+            return false;
         }
         return false;
     }
 
     public Boolean deleteFans(String fansName){
         Fans fans = getFans(fansName);
+        if (fans==null){
+            return false;
+        }
         int isOk = fansMapper.deleteFans(fans);
         if (isOk == 1){
             return true;
@@ -55,10 +67,14 @@ public class FansServiceImpl {
     }
 
     private Fans getFans(String fansName){
+        User userAuth = authDetails.getUser();
+        if (userAuth==null){
+            return null;
+        }
         User user = userMapper.selectUserByName(fansName);
         Fans fans = new Fans();
         fans.setId(IdUtil.fastSimpleUUID());
-        fans.setUserId(authDetails.getUser().getId());
+        fans.setUserId(userAuth.getId());
         fans.setFollower(user.getId());
         fans.setCreateDate(new Date());
